Do not use GitHub api to retrieve latest version
It was reported on #2 that since github api is rate limited, on some
circumstances the call may fail. To prevent any possible issues we
switch to a version constant when generating the download page for
now. On the future we could automate the retrieval of latest
stable version when a new release is published on main repo.